[Gold I] Title: K번째 수, Time: 124 ms, Memory: 14320 KB -BaekjoonHub

+++ 백준/Gold/1300. K번째 수/K번째 수.java
@@ -5,16 +5,16 @@ import java.io.InputStreamReader;
 class Main{
 	public static void main(String[] args)throws Exception{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-		long N		= Integer.parseInt(br.readLine());	// 배열크기 N (1<=십만)
-		long K		= Integer.parseInt(br.readLine());	// k(십억 <= 백억)
-		long res	= 0;								// 결과
-		long s		= 1;								// 탐색 시작값
-		long e		= K;								// 탐색 종료값, 특정 수 X는 K를 넘을 수 없다.
+		int N		= Integer.parseInt(br.readLine());	// 배열크기 N (1<=십만)
+		int K		= Integer.parseInt(br.readLine());	// k(십억 <= 백억)
+		int res		= 0;								// 결과
+		int s		= 1;								// 탐색 시작값
+		int e		= K;								// 탐색 종료값, 특정 수 X는 K를 넘을 수 없다.
 		
 		while(s <= e)
 		{
-			long mid = (s + e) >> 1;	// mid보다 작은 숫자의 개수를 찾는다.
-			long cnt = 0;
+			int mid = (s + e) >> 1;	// mid보다 작은 숫자의 개수를 찾는다.
+			int cnt = 0;
 			for(int i=1; i<=N; i++)		// 1부터 N단까지 mid로 나누었을 때 개수
 				cnt += Math.min(mid / i, N);
